optimized code for deployment

+++ client/src/components/post/PostHeading.jsx
@@ -43,19 +43,14 @@ class PostHeading extends Component {
         user : localStorage.getItem('username')
     }).then(() => {
         this.props.deletePost(this.props.postID)
-        //add delete to firebase
-        /////
-        /////
-        /////
-        ////
-        /////
-        console.log(this.props.image_uuid)
-        storage.ref().child('posts_pictures/' + this.props.image_uuid).delete()
-        .then(() => {
-        console.log('old pp deleted successfully')
-        }).catch(() => {
-        console.log('old pp could not be deleted')
-        })
+        if(this.props.image_uuid){
+            storage.ref().child('posts_pictures/' + this.props.image_uuid).delete()
+            .then(() => {
+            console.log('old pp deleted successfully')
+            }).catch(() => {
+            console.log('old pp could not be deleted')
+            })
+        }
     })
   }
   render() {
